Cleaned up App.js
Created a separate drag and drop folder
